Remove obsolete InlineBuffer Array backend
Repurpose functional tests for any chosen backend of `Array` in
Artichoke.
