+++ artichoke-backend/src/extn/core/array/array_functional_test.rb
